requirements.txt
python-telegram-bot[job-queue]==21.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
